+++ inst/examples/atomic_typecast.R
@@ -1,0 +1,29 @@
+
+# matrix example ====
+x <- matrix(rnorm(30), ncol = 5)
+colnames(x) <- month.name[1:5]
+rownames(x) <- month.abb[1:6]
+names(x) <- c(letters[1:20], LETTERS[1:10])
+print(x)
+
+as_bool(x)
+as_int(x)
+as_dbl(x)
+as_chr(x)
+as_cplx(x)
+as_raw(x)
+
+
+################################################################################
+
+# factor example ====
+x <- factor(month.abb, levels = month.abb)
+names(x) <- month.name
+print(x)
+
+as_bool(as_int(x) > 6)
+as_int(x)
+as_dbl(x)
+as_chr(x)
+as_cplx(x)
+as_raw(x)