+++ tests/tinytest.R
@@ -5,7 +5,7 @@ if ( requireNamespace("tinytest", quietly=TRUE) ){
   tinytest::test_package("tinycodet", testdir = file.path("tinytest", "safer"))
   tinytest::test_package("tinycodet", testdir = file.path("tinytest", "import"))
   tinytest::test_package("tinycodet", testdir = file.path("tinytest", "strings"))
-  tinytest::test_package("tinycodet", testdir = file.path("tinytest", "regular"))
+  tinytest::test_package("tinycodet", testdir = file.path("tinytest", "dry"))
 
   # perform special tests on fake packages:
   tinytest::test_package("tinycodet", testdir = file.path("tinytest", "special"))